requirements.txt
alabaster==0.7.13
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
Babel==2.12.1
botocore==1.31.10
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cssselect==1.2.0
defusedxml==0.7.1
distro==1.8.0
Django==3.2.16
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-dynamic-fixture==3.1.2
django-environ==0.10.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-simple-history==3.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.7
environ==1.0
exceptiongroup==1.1.1
gunicorn==20.1.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.15.0
PyJWT==2.6.0
pypng==0.20220715.0
pytest==7.3.0
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
requests==2.28.2
requests-oauthlib==1.3.1
roman==4.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
Wand==0.6.11
zipp==3.15.0
